DFS (RDMA) on byte-addressable system

+++ src/dfs/DFS.java
@@ -6,6 +6,6 @@ public interface DFS {
 
     int getLen(int inode);
     void setLen(int inode, int len); // sets length of file
-    int get(int inode, int[] buffer, int off, int len); // gets range
-    int put(int inode, int[] buffer, int off, int len); // puts range
+    int get(int inode, byte[] buffer, int off, int len); // gets range
+    int put(int inode, byte[] buffer, int off, int len); // puts range
 }
